Closes #14. Ready for check 0
