+++ holberton.h
@@ -14,5 +14,6 @@ typedef struct printer
 
 
 int _putchar(char c);
+int _printf(const char *format, ...);
 
 #endif
